created base for ML client. Additional changes in description
Notes: events are generated and placed in a folder named ics-client/events. That file is then upplaced in the DB. perhaps we can do this directly instead, that way we do not have to save the file. Also, this code does not use any LM at the moment. It uses placeholder information. I was thinking we could use llama-cpp-python. Additional changes: altered docker-compose to include network for the mongodb contianer as well as added the client container.